Move loginUser to functions
